backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-socketio[asgi]==5.11.1
pydantic==2.6.4
aiofiles==23.2.1
